# frontend/mrsexam-ui/root/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "mrsexam-backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/FormPrathamik": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "FormPrathamik"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormPrathamik"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormPrathamik"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FormPrathamik"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormPrathamik/paged": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sessionNo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormPrathamik/{id}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormPrathamik"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormPrathamik"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FormPrathamik"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormPrathamik/user/{userId}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormPrathamik/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormPrathamik/name/{name}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormPrathamik/session/{sessionNo}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "sessionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormPrathamik/region/{regionNo}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FormPrathamik": {
        "required": [
          "srNo"
        ],
        "type": "object",
        "properties": {
          "srNo": {
            "type": "integer",
            "format": "int32"
          },
          "rollNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "regionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 60,
            "type": "string",
            "nullable": true
          },
          "sessionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gender": {
            "maxLength": 1,
            "type": "string",
            "nullable": true
          },
          "formCentre": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "examCentre": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "marks1": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "totMarks": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "grade": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "undFlag": {
            "maxLength": 1,
            "type": "string",
            "nullable": true
          },
          "undClass": {
            "maxLength": 6,
            "type": "string",
            "nullable": true
          },
          "specialRank": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "result": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "dummyCentreCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "resgenRun": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "userId": {
            "maxLength": 15,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}